Explicitly specifies runtime dependency in setup.py
This patch decouples setup.py and requirements.txt. While linking the two
together is convenient, this code uses an interface exposed by pip which
is not guaranteed to be stable, and broke starting in pip 6. Since 511-transit
only has a single dependency at this time, it might be more desirable to
target newer versions of pip than to worry about duplicating dependency lists.

This patch fixes setup.py in newer pip versions (6.0 and newer).

More info:

* https://stackoverflow.com/questions/27869152/heroku-typeerror-parse-requirements-missing-1-required-keyword-argument-ses
* https://caremad.io/2013/07/setup-vs-requirement/